refactor: 로또발행 기능 서비스클래스로 변경
